# Add spinlocking for core methods, to support threading (alt to #139)