--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,11 +1,13 @@
 import ProgressMeter
 using Test
 
-# include("core.jl")
-# include("test.jl")
-# include("test_showvalues.jl")
+include("core.jl")
+include("test.jl")
+include("test_showvalues.jl")
 include("test_map.jl")
-# include("test_float.jl")
+include("test_float.jl")
+include("test_threads.jl")
+
 
 println("")
 println("All tests complete")