Trying out @KristofferC's suggestion in #139 of just spinlocking by default.

All tests pass locally for me!

```julia
n = 100
p = ProgressMeter.Progress(n)
Threads.@threads for i = 1:(n)
    sleep(0.1)
    ProgressMeter.next!(p)
end
```

I didn't figure out getting `@showprogress` working, but I assume it would follow something like intercepting `@threads` and passing to a custom `showprogressthreads`?