@ianshmean 2019-11-03
More tests. Seems ok on performance.

Master (no spinlocker):
```julia
using BenchmarkTools
@btime foo(5000) # 6.244 s (35208 allocations: 849.50 KiB)
@btime fooprogress(5000) # 6.164 s (42883 allocations: 1.42 MiB)
```

This PR (spinlock everything):
```julia
using BenchmarkTools
@btime foo(5000) # 6.043 s (35991 allocations: 897.31 KiB)
@btime fooprogress(5000) # 5.605 s (47627 allocations: 1.47 MiB)

# @threads
@btime bar(5000) # 833.957 ms (35252 allocations: 853.72 KiB)
@btime barprogress(5000) # 833.926 ms (45655 allocations: 1.33 MiB)
```

```julia
using ProgressMeter
function foo(n)
    for _ = 1:n
        sleep(0.00001)
    end
end
function fooprogress(n)
    p = Progress(n)
    for _ = 1:n
        sleep(0.00001)
        next!(p)
    end
end

function bar(n)
    Threads.@threads for _ = 1:n
        sleep(0.00001)
    end
end
function barprogress(n)
    p = Progress(n)
    Threads.@threads for _ = 1:n
        sleep(0.00001)
        next!(p)
    end
end
```