@timholy 2019-11-03
Awesome. I hate making more requests, but can you insert this into the `.travis.yml`?

```sh
script:
  - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
  - julia -e 'if VERSION < v"0.7-"; Pkg.clone(pwd()); else; import Pkg; Pkg.add(Pkg.PackageSpec(path=pwd())); end; Pkg.build("ProgressMeter")'
  - JULIA_NUM_THREADS=2 julia -e 'Pkg.test("ProgressMeter"; coverage=true)'
```

(Untested, but hopefully it will work.)