@KristofferC 2019-11-03
I think you could also make a `matrix` with env variables and the test script is run on each entry (as in https://github.com/JuliaInterop/ZMQBuilder/blob/b1fdcfbde2dceda1af78205bf8a3d81fe4fa6a08/.travis.yml#L19)