@ianshmean 2019-11-03
Looks like there's thread print issues <1.2 (and a problem on MacOS nightly..)

@KristofferC I didn't quite understand what you meant in your last suggestion