@KristofferC 2019-11-03
Just suggested an easier way to run the tests with one and many threads than to change it in the test script.