@ianshmean 2019-11-04
@KristofferC Got it. Nice and tidy. thanks

Looking at test results, is a compat of `julia = 1.2` acceptable for this package? The PR doesn't break previous usage, just when multi-threading, which seems like a lower-level julia issue