@timholy 2019-11-11
Sorry for the long delay, I have been focusing on closing a bunch of Revise bugs and limitations.

Testing at #142 whether we can make the threading tests conditional. I am basing it off https://github.com/travis-ci/travis-ci/issues/7149#issuecomment-335825605, though I don't pretend to really understand Travis' yml format.