@ianshmean 2019-11-11
@timholy I just tweaked travis on this PR to exclude specific jobs from the results of the build matrix, based on https://docs.travis-ci.com/user/build-matrix/#excluding-jobs

Seems like it worked!