@ianshmean 2019-11-11
Oh, and while travis doesn't look like it triggered on the github checks box, it actually did. Must be a github bug?
https://travis-ci.org/timholy/ProgressMeter.jl/builds/610452866?utm_medium=notification&utm_source=github_status