@ianshmean 2019-11-11
I can't see why the linux 1.2 failed.. I couldn't see a source of the error. That might just need a restart