@karajan9 2019-11-13
> ```julia
> sleep(0.00001)
> ```

Since I stumbled over this recently: `sleep` only goes down to `sleep(0.001)` as you can also see in the resulting time. But I think this shouldn't affect the results of the test.